refactor: move getDateTimeTerms fn to lib utils
