+++ components/Package/Package.test.js
@@ -2,9 +2,9 @@ import { mount } from 'enzyme';
 import Router from 'next/router';
 import wait from 'waait';
 import { normalizeItem } from 'lib/elastic/parser';
-import { getPluralStringOrNot, suppressActWarning } from 'lib/utils';
+import { getDateTimeTerms, getPluralStringOrNot, suppressActWarning } from 'lib/utils';
 import DownloadPkgFiles from 'components/admin/download/DownloadPkgFiles/DownloadPkgFiles';
-import { getDateTimeTerms, normalizeDocumentItemByAPI } from './utils';
+import { normalizeDocumentItemByAPI } from './utils';
 import Package from './Package';
 import { packageItem } from './packageElasticMock';
 import { packageItemGraph } from './packageGraphMock';
