+++ lib/utils.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-mixed-operators */
 import config from 'config';
+import moment from 'moment';
 import getConfig from 'next/config';
 import forEach from 'lodash/forEach';
 import htmlParser from 'react-markdown/plugins/html-parser';
@@ -674,3 +675,24 @@ export const parseToParagraphs = text => {
   }
 };
 
+/*
+ * Format date/time for package, document cards display
+ *
+ */
+export const getDateTimeTerms = ( createdAt, updatedAt, format ) => {
+  const isDocUpdated = updatedAt > createdAt;
+
+  const dateTimeStamp = isDocUpdated ? updatedAt : createdAt;
+  // const label = isDocUpdated ? 'Updated' : 'Created';
+  // published and modified are equal, so setting label to 'Updated
+  // TO DO: Add createdAt to transform
+  const label = 'Updated';
+
+  return [
+    {
+      definition: <time dateTime={ dateTimeStamp }>{ `${moment( dateTimeStamp ).format( format )}` }</time>,
+      displayName: label,
+      name: label,
+    },
+  ];
+};
